Moved new enums to io.realm
